mv scaling namespace to deep_tau
